rocketwatch/requirements.txt
psutil==7.2.2
python_i18n==0.3.9
web3>=7.0.0,<8.0.0
humanize==4.15.0
termplotlib==0.3.9
cachetools==7.0.3
bidict==0.23.1
uptime==3.0.1
discord.py==2.7.1
pydantic>=2.0.0,<3.0.0
pytz==2026.1.post1
matplotlib==3.10.8
inflect==7.5.0
colorama==0.4.6
seaborn==0.13.2
etherscan_labels @ git+https://github.com/haloooloolo/etherscan-labels
cronitor==4.9.0
retry-async==0.1.4
dice==4.0.0
regex==2026.2.28
tiktoken==0.12.0
anthropic==0.84.0
pymongo==4.16.0
graphql_query==1.4.0
pillow==12.1.1
aiohttp==3.13.3
aiocache==0.12.3
numpy==2.4.2
beautifulsoup4==4.14.3
eth-typing==5.2.1
hexbytes==1.3.1
eth-utils==5.3.1
tabulate==0.10.0
anyascii==0.3.3
